+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'untidy-board'
+rootProject.name = 'bamboo-forest'
