+++ src/main/resources/application.properties
@@ -1,1 +1,4 @@
-
+spring.datasource.url=jdbc:mysql://localhost/untidy?characterEncoding=UTF-8&serverTimezone=UTC
+spring.datasource.username=root
+spring.datasource.password=w30t87ch
+spring.datasource.driver-class-name=com.mysql.jdbc.Driver
