+++ src/test/java/com/minjae/bambooforest/UntidyBoardApplicationTests.java
@@ -1,4 +1,4 @@
-package com.minjae.untidyboard;
+package com.minjae.bambooforest;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
